// src/generated/models/MatchMode.generated.ts
/* eslint-disable */
/**
 * This file was automatically generated.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file.
 *
 * =>  @see ./src/schemas/models/MatchMode.yml
 *
 * and run "npm run schema-to-interface" or "npm run generate" to regenerate this file.
 */

import Ajv, {ValidateFunction} from "ajv";
import addFormats from "ajv-formats";

export const schemaForMatchMode = {$id: "MatchMode.yml", type: "string", enum: ["any", "all"]};

export function validateMatchMode(o: object): {isValid: boolean; validate: ValidateFunction} {
  const ajv = new Ajv();
  addFormats(ajv);
  ajv.addKeyword("example");

  const validate = ajv.compile(schemaForMatchMode);
  return {isValid: validate(o), validate: validate};
}

export type MatchMode = "any" | "all";